todo-web-app/backend/requirements.txt
# Core web/backend
fastapi==0.115.0
uvicorn==0.32.0
starlette==0.38.6

# Database
SQLAlchemy==2.0.45
sqlmodel==0.0.20
asyncpg==0.31.0
psycopg2-binary==2.9.11
alembic==1.13.3

# Security
bcrypt==5.0.0
passlib==1.7.4
python-jose==3.3.0
cryptography==46.0.3

# HTTP clients
httpx==0.28.1
anyio==4.12.0
aiohttp==3.13.2

# Environment variables
python-dotenv==1.2.1

# Data handling
pydantic==2.12.5
pydantic-settings==2.12.0
orjson==3.11.5
numpy==2.3.5
openpyxl==3.1.5
lxml==6.0.2

# Optional utilities
tqdm==4.67.1
requests==2.32.5
typing_extensions==4.15.0

# AI/Chatbot
openai>=1.3.0

# Remove Windows-only and unnecessary packages:
# pywin32, git-filter-repo, huggingface, langchain, cohere, greenlet, etc.
# Those are either for local development or not needed on Render.
